Serialisation formats namespace
