+++ Sources/Scout/Models/PathExplorer/DataFormat.swift
@@ -1,0 +1,9 @@
+//
+// Scout
+// Copyright (c) Alexis Bridoux 2020
+// MIT license, see LICENSE file for details
+
+/// Unique identifier of a data format
+public enum DataFormat: String, CaseIterable {
+    case json, plist, xml, yaml
+}
